Move transfer export queries behind useQueries and fix hero preload.
Keep request parsing in the transfer service, put Drizzle I/O in db/queries, and use Nuxt Image preload with fetchPriority instead of the invalid priority prop.
