+++ apps/cms/server/api/transfer/articles.get.ts
@@ -1,14 +1,11 @@
 import { requireApiKey } from '../../../utils/api-key-auth'
-import { useDb } from '../../../utils/db'
-import {
-  createTransferExportQueries,
-  parseTransferPage,
-} from '../../../services/transfer-export'
+import { useQueries } from '../../../utils/db'
+import { parseTransferPage } from '../../../services/transfer-export'
 
 export default defineEventHandler(async (event) => {
   await requireApiKey(event, 'articles')
   const { limit, cursor } = parseTransferPage(getQuery(event) as Record<string, unknown>)
-  const page = await createTransferExportQueries(useDb(event)).exportArticlesPage({
+  const page = await useQueries(event).transferExport.exportArticlesPage({
     cursor,
     limit,
   })
